Extract strategies into proper module
