+++ tests/test_funcs.py
@@ -18,7 +18,7 @@ from attr._compat import TYPE
 from attr.exceptions import AttrsAttributeNotFoundError
 from attr.validators import instance_of
 
-from .utils import nested_classes, simple_classes
+from .strategies import nested_classes, simple_classes
 
 
 MAPPING_TYPES = (dict, OrderedDict)
